added friends component. separate friends functionality from dashboard.
