+++ pages/nav/NavBar.js
@@ -12,6 +12,7 @@ const NavBar = () => {
                 <>
                     <p className="subtitle">Hello, {JSON.parse(JSON.stringify(user.username))}</p>
                     <NavBarItem link="/account/Dashboard" title="Dashboard" />
+                    <NavBarItem link="/account/Friends" title="Friends" />
                     <NavBarItem link="/api/account/logout" title="Log out" />
                 </>
             ) : (
